Relayout modular app at repository root
