+++ Console-ComputationalVision/application/send_raw_command.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ..domain.motion.gcode_sender import GcodeSender
+from domain.motion.gcode_sender import GcodeSender
 from .send_coordinates import CommandDispatcher, CommandRequest
