+++ Console-ComputationalVision/domain/camera/camera.py
