+++ Console-ComputationalVision/domain/camera/camera_repository.py
