+++ Console-ComputationalVision/domain/events.py
