+++ Console-ComputationalVision/domain/vision/detector.py
