+++ Console-ComputationalVision/domain/vision/model.py
