+++ Console-ComputationalVision/infrastructure/logging_sink.py
