+++ Console-ComputationalVision/presentation/widgets/python_logs.py
