+++ Console-ComputationalVision/shared/errors.py
