+++ Console-ComputationalVision/shared/utils.py
