+++ Console-ComputationalVision/tests/integration/test_system_smoke.py
@@ -5,19 +5,19 @@ import time
 
 import numpy as np
 
-from ...application.poll_current_position import PollCurrentPositionUseCase, PositionPoller
-from ...application.send_coordinates import CommandDispatcher, SendCoordinatesUseCase
-from ...application.start_detection import DetectionController, StartDetectionUseCase, StopDetectionUseCase
-from ...application.update_settings import VisionSettingsStore
-from ...domain.camera.camera import Camera, Frame, Resolution
-from ...domain.camera.camera_repository import CameraRepository
-from ...domain.events import DetectionProduced
-from ...domain.motion.gcode_sender import CommandAck, GcodeSender
-from ...domain.motion.position import Feedrate, Position
-from ...domain.settings.settings import VisionSettings
-from ...domain.vision.detector import Detector
-from ...domain.vision.model import BoundingBox, InferenceResult
-from ...shared.bus import EventBus
+from application.poll_current_position import PollCurrentPositionUseCase, PositionPoller
+from application.send_coordinates import CommandDispatcher, SendCoordinatesUseCase
+from application.start_detection import DetectionController, StartDetectionUseCase, StopDetectionUseCase
+from application.update_settings import VisionSettingsStore
+from domain.camera.camera import Camera, Frame, Resolution
+from domain.camera.camera_repository import CameraRepository
+from domain.events import DetectionProduced
+from domain.motion.gcode_sender import CommandAck, GcodeSender
+from domain.motion.position import Feedrate, Position
+from domain.settings.settings import VisionSettings
+from domain.vision.detector import Detector
+from domain.vision.model import BoundingBox, InferenceResult
+from shared.bus import EventBus
 
 
 class FakeDetectorFactory:
